feat(replies): implement reply validation and middleware
Add reply validators for params and content
Create comment middleware for reply routes
Update reply controllers to use validators and middleware
Improve error handling and response messages

+++ src/app/views/video/comments/delComment.js
@@ -8,8 +8,9 @@ const delComment = async (req, res) => {
     const deletedComment = await Comment.findOneAndDelete({ _id: commentID, user: userID })
     if (!deletedComment) throw new Error("COMMENT_NOT_FOUND_OR_UNAUTHORIZED")
     await Promise.all([
-      Comment.deleteMany({ parent: commentID }),
-      Video.updateOne({ _id: videoID }, { $pull: { comments: commentID } })
+      deletedComment.parent && Comment.updateOne({ _id: deletedComment.parent }, { $pull: { replies: deletedComment._id } } ),
+      !deletedComment.parent && Comment.deleteMany({ parent: commentID }),
+      !deletedComment.parent && Video.updateOne({ _id: videoID }, { $pull: { comments: commentID } })
     ])
     return res.status(204).send()
   } catch (error) {
